+++ webrender_api/src/lib.rs
@@ -52,6 +52,7 @@ mod display_list;
 mod font;
 mod gradient_builder;
 mod image;
+mod fast_transform;
 mod tile_pool;
 pub mod units;
 
@@ -63,6 +64,7 @@ pub use crate::font::*;
 pub use crate::gradient_builder::*;
 pub use crate::image::*;
 pub use crate::tile_pool::*;
+pub use crate::fast_transform::*;
 
 use crate::units::*;
 use crate::channel::Receiver;
